Add a Makefile for common tasks
This is a super simple modern go project that uses gomodules, so the
commands to run tests and compile the provider are easy to remember.

Still, it's nice to wrap them in make targets so folks less familiar
with go can get up and running with development. I also expected we
might end up with some more complex make targets when building a github
release pipeline.