Add acceptance tests for the agent_token resource
These are the first acceptance tests for the provider, so they come with
a bunch of supporting infrastructure:

* a new Make target
* Some instructions in the README
* a new pipeline step that can only run on the main branch
  * if a third party PR removes this restriction, the step will run but
    fail because it'll run on an agent with no access to the required
    credentials in SSM ParameterStore

The CI setup here is imperfect. Pull requests won't run the acceptance
tests and there's a risk we'll only notice a failure in the acceptance
tests after merging. For now we'll have to live with that (maybe by
manually running the tests ourselves prior to merging?) while we discuss
a better approach.