add some acceptance tests for the pipeline resource
This resource had no acceptance tests, and we were nervous about the
risk of regressions.

This adds a few initial test cases:

* creating a pipeline with only the required attributes set
* creating a pipeline with all attributes set
* updating a pipeline
* importing an existing pipeline

Pipeline deletion is tested implicitly when the terraform testing sdk
cleans up after a test.

We assume there's potential for more tests that exercise specific
combinations (for example, exposing this teams related bug in pipeline
creation #84), but this is a good start.

Co-authored-by: Chloe <chloe@buildkite.com>