Add acceptance tests for teams
This is one of the resources that does not yet have acceptance tests
associated with it. In order to guard against regressions, I have added
the following tests:

* creating a team (with only the required attributes)
* updating a team (changing its name, description and visibility)
* importing a team

During this work I discovered an error in importing teams. In order to
resolve this, I added the DefaultMemberRole value which was present in
the schema to the TeamNode struct, and added a mapping to updateTeam, so
this PR doubles as a bugfix for import.