Add a pipeline data source
This adds a data source for fetching the properties of a single
pipeline. The pipeline is specified using a slug, available in the
pipeline URL on buildkite.com.

This could be particularly useful for folks who want to lookup the
webhook URL and set it on a GitHub repo (using the  GitHub terraform
provider).

Closes #93