stop using deprecated part of the terraform sdk API
While working on #106 I noticed that v2 of the terraform sdk has
deprecated the Create/Read/Update/Delete properties of the
schema.Resource struct in favour of new properties that support passing
context.

I thought I'd bite the bullet and just make the change everywhere so all
current resources are a good model for adding new resources in the
future.

Note that we only updated to v2 of the sdk in #89 ~3 weeks ago, so this
deprecation only appeared for us recently.

I haven't made an effort to use the new context instances to improve the
way timeouts are handled when we query the graphql API. That could be
worth exploring in the future.