Add concurrency limit to acceptance test steps
Because we use named resources on Buildkite in our acceptance tests,
when multiple steps of this type run in parallel (e.g. if we have
multiple merges at the same time) it can cause the tests to fail when
they try to create resources with the same tokens in the test
organisation. In order to fix this, I've set the concurrency limit to 1
on a new "terraform-acceptance-tests" concurrency group.

This won't fix the scenario of a developer running the acceptance tests
locally at the same time as a build happens to run them too, but the
scenario is relatively unlikely and we can just retry it. If this
repository gets significantly more active we can revisit this and try to
fix it by giving resources unique per-run names.