Adjust pipeline creation so it works for non-admins
For Buildkite organizations with teams enabled, it's possible for
non-admins to create pipelines within their own teams.

However, it only works if the team is included in the pipelineCreate()
mutation. For a non-admin graphql token, pipelineCreate() will return an
error if the input includes no teams.

The existing code was calling pipelineCreate() with no teams, and then
following up with a teamPipelineCreate() mutation to apply the teams. We
never got to the follow up though, because the initial pipelineCreate()
was failing.

This merges the two steps so the initial pipelineCreate() includes
teams. Once the pipeline has been created, team management is still done
independently of the rest of the pipeline. That should be fine though,
it's only pipelineCreate() that requires teams to be include2.
Non-admin graphql tokens can call pipelineUpdate() without teams just
fine.

Fixes #84