adjust cron line we use in pipeline_schedule update test
pipeline schedules are not permitted to run more frequently than every
10 minutes.

This test was updating the cronline to "* 0 * * *", which is "every
minute during hour 0". Apparently our server side validation only
detects that as an error when the update runs during UTC hour 0, so this
test fails when run during UTC hour 0 [1].

The test now updates the cronline to "daily at 1:00 UTC".

[1] https://buildkite.com/buildkite/terraform-provider-buildkite-main/builds/43#c6ca85d2-9311-43c8-91a0-9290460242c0