Add build step for subset of acctests that will work with a non-admin API token
In PR #112 we fixed the pipeline resource so it can be created and
updated by non-admin API tokens.

To prevent regressions, this adds a new acceptance test step to the
build pipeline that will run only the subset of acceptance tests that
will work with a non-admin API token.

Non-admins can manage pipelines and pipeline schedules, but only if the
pipeline belongs to a team where the user is a team maintainer.

Before this we had no acctests with teams declared on a pipeline, so
none of them would work for non-admins. This has added a couple of basic
tests that use teams.

The new tests will also run in the normal acctest step (using an admin
token), which is a nice bonus.