Return a better user visible error when pipeline references an unrecognised team slug
In #117 @keith reported that we return a particularly unhelpful error to
users when a pipeline references a team slug that doesn't exist:

    resource "buildkite_pipeline" "repo2" {
        name = "repo2"
        repository = "git@github.com:org/repo2"
        steps = file("./steps.yml")

        team {
            slug = "not valid"
            access_level = "READ_ONLY"
        }
    }

The error was:

> Error: No ID was supplied

I'm not 100% sure this error handling is where the error @keith saw was
from, but it seems pretty likely. Even if it's not, this seems like a
good change to make.

The Printf() is only visible to users when debug mode is on, which will
almost never be true. Moving that message to an error that will be
displayed to the user (and can be searched in this codebase while
debugging) seems like a win.